Added workout validation so workout is not in the future and needs to be older than an hour
Reworked FitBit api and was able to add refresh method on the model.

Added FitBit authentication buttons to the show workouts page.

Created a custom route to get FitBit information without needing to enter edit and save.

Fixed new form to adjust for the new default attribute on a the Memberships model.

Created scopes on the models in order to pull the default gym for the user.

Added FitBit authentication option to the nav bar.